Refactored recoverable connection into individual types, one per category
