+++ sdk/eventhubs/azure_messaging_eventhubs/src/common/recoverable/mod.rs
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft Corporation. All Rights reserved
+// Licensed under the MIT license.
+
+mod cbs;
+mod connection;
+mod management;
+mod receiver;
+mod sender;
+
+pub(crate) use connection::RecoverableConnection;
+pub(crate) use sender::RecoverableSender;
